requirements.txt
numpy==1.26.4
pandas==2.1.4
scipy==1.11.3
pdfplumber==0.9.0
PyMuPDF==1.21.1
camelot-py==0.9.0
opencv-python==4.8.1.78
transformers==4.35.2
sentence-transformers==2.7.0
scikit-learn==1.3.0
modAL
ruamel.yaml==0.17.32
joblib==1.3.2
pytest==7.4.3
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.2
eel
pyinstaller==6.2.0
Pillow==10.0.1
openpyxl==3.1.2

# PaddleOCR dependencies
paddlepaddle==2.6.1
paddleocr>=2.8.1

# Torch CPU versions with official stable repo
# Установка torch через pip с индексом PyTorch:
# pip install torch==2.4.0 torchvision==0.19.0 torchaudio==2.4.0 --index-url https://download.pytorch.org/whl/cpu
# Или для GPU версии:
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118
